requirements.txt
# macOS pip 格式依赖文件
blinker==1.6.2
bottleneck==1.3.7
Brotli==1.0.9
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
contourpy==1.0.5
cryptography==41.0.3
cycler==0.11.0
flask==3.0.3
fonttools==4.51.0
idna==3.4
importlib-metadata==7.0.1
importlib_resources==6.1.1
itsdangerous==2.2.0
jieba==0.42.1
jinja2==3.1.4
joblib==1.2.0
kiwisolver==1.4.4
markupsafe==2.1.3
matplotlib==3.7.2
numexpr==2.8.4
numpy==1.24.3
packaging==23.1
pandas==2.0.3
pillow==10.3.0
pooch==1.7.0
pycparser==2.21
pymysql==1.0.2
pyopenssl==23.2.0
pyparsing==3.0.9
pysocks==1.7.1
python-dateutil==2.9.0
tzdata==2023.3
pytz==2024.1
requests==2.31.0
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
snownlp==0.12.3
threadpoolctl==2.2.0
urllib3==1.26.18
werkzeug==3.0.3
wordcloud==1.9.3
zipp==3.17.0
DBUtils==3.0.2
bleach==6.1.0
backports.zoneinfo==0.2.1;python_version<"3.9"
apscheduler==3.10.4
python-dotenv==1.0.1
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
pmdarima==2.0.4
openai==1.6.1
anthropic==0.8.1
transformers==4.37.2
sqlalchemy==2.0.27
Flask-Limiter==3.5.0
redis==5.0.1
argon2-cffi==23.1.0
websockets==12.0
aiohttp==3.9.1
ratelimit==2.2.1
tenacity==8.2.3
psycopg2-binary==2.9.9
pymongo==4.6.1
httpx~=0.28.1
backports~=1.1
tqdm~=4.67.1
bs4~=4.12.3
beautifulsoup4~=4.12.3
cachetools~=5.3.3
gensim~=4.3.3
flask-sock==0.7.0